services/kill.py: Allow admins to kill any run

+++ src/teuthology_api/services/kill.py
@@ -4,18 +4,20 @@ import subprocess
 
 from fastapi import HTTPException, Request
 
-from teuthology_api.services.helpers import get_username, get_run_details
+from teuthology_api.services.helpers import get_username, get_run_details, isAdmin
 
 
 TEUTHOLOGY_PATH = os.getenv("TEUTHOLOGY_PATH")
+ADMIN_TEAM = os.getenv("ADMIN_TEAM")
+
 log = logging.getLogger(__name__)
 
 
-def run(args, send_logs: bool, access_token: str, request: Request):
+async def run(args, send_logs: bool, token: dict, request: Request):
     """
     Kill running teuthology jobs.
     """
-    if not access_token:
+    if not token:
         log.error("access_token empty, user probably is not logged in.")
         raise HTTPException(
             status_code=401,
@@ -23,41 +25,51 @@ def run(args, send_logs: bool, access_token: str, request: Request):
             headers={"WWW-Authenticate": "Bearer"},
         )
     username = get_username(request)
-    run_name = args.get("--run")
+    run_name = args.get("--run", "")
     if run_name:
         run_details = get_run_details(run_name)
-        run_owner = run_details.get("user")
+        jobs_details = run_details.get("jobs", [])
+        if jobs_details:
+            run_owner = jobs_details[0].get("owner", "")
     else:
         log.error("teuthology-kill is missing --run")
         raise HTTPException(status_code=400, detail="--run is a required argument")
-    # TODO if user has admin priviledge, then they can kill any run/job.
-    if run_owner.lower() != username.lower():
-        log.error(
-            "%s doesn't have permission to kill a job scheduled by: %s",
-            username,
-            run_owner,
-        )
-        raise HTTPException(
-            status_code=401, detail="You don't have permission to kill this run/job"
-        )
+
+    if (run_owner.lower() != username.lower()) and (
+        run_owner.lower() != f"scheduled_{username.lower()}@teuthology"
+    ):
+        isUserAdmin = await isAdmin(username, token.get("access_token"))
+        if not isUserAdmin:
+            log.error(
+                "%s doesn't have permission to kill a job scheduled by: %s",
+                username,
+                run_owner,
+            )
+            raise HTTPException(
+                status_code=401, detail="You don't have permission to kill this run/job"
+            )
+        log.info("Killing with admin privileges")
     try:
         kill_cmd = [f"{TEUTHOLOGY_PATH}/virtualenv/bin/teuthology-kill"]
         for flag, flag_value in args.items():
-            if isinstance(flag_value, bool):
-                flag_value = int(flag_value)
-            kill_cmd += [flag, str(flag_value)]
+            if isinstance(flag_value, bool):  # check for --preserve-queues
+                if flag_value == True:
+                    kill_cmd += [flag]
+            else:
+                kill_cmd += [flag, str(flag_value)]
         log.info(kill_cmd)
         proc = subprocess.Popen(
             kill_cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
         stdout, stderr = proc.communicate()
         returncode = proc.wait(timeout=120)
-        log.info(stdout)
+        output_logs = stdout.decode()
+        log.info(output_logs)
         if returncode != 0:
-            raise Exception(stdout)
+            raise Exception(output_logs)
         if send_logs:
             return {"kill": "success", "logs": stdout}
         return {"kill": "success"}
     except Exception as exc:
         log.error("teuthology-kill command failed with the error: %s", repr(exc))
-        raise HTTPException(status_code=500, detail=repr(exc)) from exc
+        raise HTTPException(status_code=500, detail=str(exc)) from exc
